Fix overlay registration and client tick events
